daemon/requirements.txt
alpaca-py>=0.37.0
pyyaml>=6.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
# sqlite3 is stdlib — no install needed
